+++ apps/everest/pages/db-cluster-details/backups/backup.context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { ScheduleModalContextType } from './backup.types';
+
+export const ScheduleModalContext = createContext<ScheduleModalContextType>({});
