+++ apps/everest/pages/db-cluster-details/backups/backup.types.ts
@@ -1,0 +1,8 @@
+export type ScheduleModalContextType = {
+  mode?: 'new' | 'edit';
+  setMode?: React.Dispatch<React.SetStateAction<'new' | 'edit'>>;
+  selectedScheduleName?: string;
+  setSelectedScheduleName?: React.Dispatch<React.SetStateAction<string>>;
+  openScheduleModal?: boolean;
+  setOpenScheduleModal?: React.Dispatch<React.SetStateAction<boolean>>;
+};
