+++ apps/everest/pages/db-cluster-details/backups/backups-list/backups-list.tsx
@@ -38,8 +38,8 @@ import { Backup, BackupStatus } from '../../../../types/backups.types';
 import { DbEngineType } from '../../../../types/dbEngines.types';
 import { BACKUP_STATUS_TO_BASE_STATUS } from './backups-list.constants';
 import { Messages } from './backups-list.messages';
-import { ScheduleModalContext } from '../scheduled-backup-modal/context/schedule-modal.context';
 import { OnDemandBackupModal } from '../on-demand-backup-modal/on-demand-backup-modal';
+import { ScheduleModalContext } from '../backup.context';
 
 export const BackupsList = () => {
   const navigate = useNavigate();
