+++ apps/everest/pages/db-cluster-details/backups/backups.tsx
@@ -13,14 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useMemo } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useDbCluster } from '../../../hooks/api/db-cluster/useDbCluster';
 import { DbEngineType } from '../../../types/dbEngines.types';
 import { BackupsList } from './backups-list/backups-list';
 import { ScheduledBackupsList } from './scheduled-backups-list/scheduled-backups-list';
-import { ScheduleModalContextProvider } from './scheduled-backup-modal/context/schedule-modal.context';
 import { ScheduledBackupModal } from './scheduled-backup-modal/scheduled-backup-modal';
+import { ScheduleModalContext } from './backup.context';
 
 export const Backups = () => {
   const { dbClusterName } = useParams();
@@ -29,18 +29,31 @@ export const Backups = () => {
     enabled: !!dbClusterName,
   });
 
+  const [mode, setMode] = useState<'new' | 'edit'>('new');
+  const [openScheduleModal, setOpenScheduleModal] = useState(false);
+  const [selectedScheduleName, setSelectedScheduleName] = useState<string>('');
+
   const dbType = useMemo(
     () => dbCluster?.spec?.engine?.type,
     [dbCluster?.spec?.engine?.type]
   );
 
   return (
-    <ScheduleModalContextProvider>
+    <ScheduleModalContext.Provider
+      value={{
+        mode,
+        setMode,
+        openScheduleModal,
+        setOpenScheduleModal,
+        selectedScheduleName,
+        setSelectedScheduleName,
+      }}
+    >
       <>
         {dbType !== DbEngineType.POSTGRESQL && <ScheduledBackupsList />}
         <BackupsList />
-        <ScheduledBackupModal />
+        {openScheduleModal && <ScheduledBackupModal />}
       </>
-    </ScheduleModalContextProvider>
+    </ScheduleModalContext.Provider>
   );
 };
